# server/user/generated-client/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: devoops@tyrean.mozmail.com
    name: Team DevOops
    url: https://github.com/AET-DevOps25/team-devoops/
  description: "This OpenAPI specification defines the endpoints, schemas, and security\
    \ mechanisms\nfor the Meet@Mensa User micro-service. "
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: ""
  title: MeetAtMensa
  version: 2.3.0
servers:
- description: ""
  url: api.meetatmensa.com
security:
- jwt-bearer: []
...
paths:
...
  /api/v2/matching/request/submit:
    post:
      description: Submit a new matching request to the Matching-Service
      operationId: post-api-v2-matching-request-submit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequestNew"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
          description: Request submitted sucessfully
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Submit matching Request
      tags:
      - Matching
      x-stoplight:
        id: 1x5ta8qguutzq
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v2/matching/matches/{user-id}:
    get:
      description: "Retrieve all matches for a user with {user-id} from the matching-service"
      operationId: get-api-v2-matching-matches-userID
      parameters:
      - $ref: "#/components/parameters/user-id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchCollection"
          description: OK
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: "Retrieve all matches for a {user-id}"
      tags:
      - Matching
      x-stoplight:
        id: nxywx9cl1qg31
      x-accepts:
      - application/json
    parameters:
    - $ref: "#/components/parameters/user-id"
  /api/v2/matching/requests/{user-id}:
    get:
      description: "Retrieve all MatchRequests for a user with {user-id} from the\
        \ matching-service"
      operationId: get-api-v2-matching-requests-userID
      parameters:
      - $ref: "#/components/parameters/user-id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequestCollection"
          description: OK
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: "Retrieve all MatchRequests for a {user-id}"
      tags:
      - Matching
      x-stoplight:
        id: wkc1ys6vkbv12
      x-accepts:
      - application/json
    parameters:
    - $ref: "#/components/parameters/user-id"
  /api/v2/matching/request/{request-id}:
    delete:
      description: "Delete MatchRequest with ID {request-id} from the system"
      operationId: delete-api-v2-matching-request-request-id
      parameters:
      - $ref: "#/components/parameters/request-id"
      responses:
        "200":
          description: Request Deleted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: "Delete MatchRequest with {request-id}"
      tags:
      - Matching
      x-stoplight:
        id: shgx14ydoaa2f
      x-accepts:
      - application/json
    parameters:
    - $ref: "#/components/parameters/request-id"
    put:
      description: "Update all information in the MatchRequest with ID {request-id}"
      operationId: put-api-v2-matching-request-request-id
      parameters:
      - $ref: "#/components/parameters/request-id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequestUpdate"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequest"
          description: Ok
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "406":
          description: MatchRequest cannot be updated since it has already been fulfilled!
        "500":
          description: Internal Server Error
      summary: "Update MatchRequest with {request-id}"
      tags:
      - Matching
      x-stoplight:
        id: f6t3aeqcvup68
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v2/matching/rsvp/{match-id}/accept:
    get:
      description: Accept invitation to a given match
      operationId: get-api-v2-matching-rsvp-match-id-accept
      parameters:
      - description: UUID associated with a given match
        explode: false
        in: path
        name: match-id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: Accept invitation to a given match
      tags:
      - Matching
      x-stoplight:
        id: am6b7xnyytanu
      x-accepts:
      - application/json
    parameters:
    - description: UUID associated with a given match
      explode: false
      in: path
      name: match-id
      required: true
      schema:
        format: uuid
        type: string
      style: simple
  /api/v2/matching/rsvp/{match-id}/reject:
    get:
      description: Reject invitation to a given match
      operationId: get-api-v2-matching-rsvp-match-id-reject
      parameters:
      - $ref: "#/components/parameters/match-id"
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          description: Internal Server Error
      summary: Reject invitation to a given match
      tags:
      - Matching
      x-stoplight:
        id: swjozsxt7gvip
      x-accepts:
      - application/json
    parameters:
    - $ref: "#/components/parameters/match-id"
...
  /api/v2/matching/demo:
    post:
      description: Submit a match request which will be immediately matched with a
        group of demo users.
      operationId: post-api-v2-matching-demo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequestNew"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
          description: OK
        "409":
          description: User already has a meeting on this day!
        "500":
          description: Internal Server Error
      summary: Create demo request
      tags:
      - Matching
      x-stoplight:
        id: nz0sz8ks59ugw
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  parameters:
    user-id:
      description: UUID associated with a given user
      explode: false
      in: path
      name: user-id
      required: true
      schema:
        format: uuid
        type: string
      style: simple
...
  responses:
    UnauthorizedError:
      description: |
        Authentication failed due to missing or invalid OAuth2 token.
    NotFoundError:
      description: |
        The requested resource was not found.
    BadRequestError:
      description: |
        The request was malformed or contained invalid parameters.
  schemas:
...
    MatchRequest:
      description: Object representing a request for matching a given user on a given
        date in the Meet@Mensa system.
      example:
        date: 2000-01-23
        preferences:
          degreePref: true
          agePref: true
          genderPref: true
        requestID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        timeslot:
        - null
        - null
        location: GARCHING
        userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: PENDING
      properties:
        requestID:
          description: The unique ID of a MatchRequest in the Meet@Mensa system.
          format: uuid
          title: requestID
          type: string
          x-stoplight:
            id: rjnh43gvsnw2a
        userID:
          description: The unique ID of a single student in the Meet@Mensa system.
          format: uuid
          title: userID
          type: string
          x-stoplight:
            id: qlawgs207p4xu
        date:
          description: The date a user would like meet@mensa to find them a match
          format: date
          type: string
          x-stoplight:
            id: hpw97fh7xj03p
        timeslot:
          items:
            $ref: "#/components/schemas/timeslot"
          type: array
          x-stoplight:
            id: 7q7ast68lvs4u
        location:
          $ref: "#/components/schemas/location"
        preferences:
          $ref: "#/components/schemas/MatchPreferences"
        status:
          $ref: "#/components/schemas/requestStatus"
      required:
      - date
      - location
      - preferences
      - requestID
      - status
      - timeslot
      - userID
      title: MatchRequest
      x-stoplight:
        id: mbhmumrcj2bd4
    MatchRequestCollection:
      description: Object representing a collection of match requests in the Meet@Mensa
        system.
      example:
        requests:
        - date: 2000-01-23
          preferences:
            degreePref: true
            agePref: true
            genderPref: true
          requestID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          timeslot:
          - null
          - null
          location: GARCHING
          userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: PENDING
        - date: 2000-01-23
          preferences:
            degreePref: true
            agePref: true
            genderPref: true
          requestID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          timeslot:
          - null
          - null
          location: GARCHING
          userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: PENDING
      properties:
        requests:
          items:
            $ref: "#/components/schemas/MatchRequest"
          type: array
          x-stoplight:
            id: wfe1ymfi052cg
      required:
      - requests
      title: MatchRequestCollection
      x-stoplight:
        id: ccqjyx1unnz6s
    MatchRequestNew:
      description: Object representing a request for matching a given user on a given
        date in the Meet@Mensa system.
      example:
        date: 2000-01-23
        preferences:
          degreePref: true
          agePref: true
          genderPref: true
        timeslot:
        - null
        - null
        location: GARCHING
        userID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        userID:
          description: The unique ID of a single student in the Meet@Mensa system.
          format: uuid
          title: userID
          type: string
          x-stoplight:
            id: qlawgs207p4xu
        date:
          description: The date a user would like meet@mensa to find them a match
          format: date
          type: string
          x-stoplight:
            id: 4p1ncij1h0rb0
        timeslot:
          items:
            $ref: "#/components/schemas/timeslot"
          type: array
          x-stoplight:
            id: nm1ht2n4e7m9o
        location:
          $ref: "#/components/schemas/location"
        preferences:
          $ref: "#/components/schemas/MatchPreferences"
      required:
      - date
      - location
      - preferences
      - timeslot
      - userID
      title: MatchRequestNew
      x-stoplight:
        id: 3k8e2egr16sfx
...
  securitySchemes:
    jwt-bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
...